Move new logging to logging.h
