+++ src/Qubic.vcxproj.filters
@@ -55,6 +55,7 @@
       <Filter>smart_contracts</Filter>
     </ClInclude>
     <ClInclude Include="system.h" />
+    <ClInclude Include="logging.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="smart_contracts">
